Work-in-progress splitting schema up into smaller files, repo organization
