+++ api/src/graphql/schema/root.js
@@ -1,0 +1,9 @@
+const { gql } = require("apollo-server");
+
+const root = gql`
+    type Query {
+        helloWorld: String
+    }
+`;
+
+module.exports = root;
