+++ api/src/graphql/schema/user.js
@@ -1,0 +1,26 @@
+const { gql } = require("apollo-server");
+
+const user = gql`
+    type Query {
+        getUser(id: ID!): User
+        getAllUsers: [User!]!
+    }
+
+    type Mutation {
+        updateUser(
+            id: ID!
+            firstName: String
+            lastName: String
+            email: String
+        ): User
+    }
+
+    type User {
+        id: ID!
+        firstName: String
+        lastName: String
+        email: String
+    }
+`;
+
+module.exports = user;
