--- api/src/schema/resume.js
@@ -1,19 +1,0 @@
-const types = `
-
-`;
-
-const queries = `
-    type Query {
-        
-    }
-`;
-
-const mutations = `
-
-`;
-
-module.exports = {
-    types,
-    queries,
-    mutations,
-};
