+++ package.json
@@ -50,7 +50,7 @@
         "jwks-rsa": "^1.7.0",
         "knex": "^0.20.15",
         "knex-cleaner": "^1.3.0",
-        "merge-graphql-schemas": "^1.7.7",
+        "merge-graphql-schemas": "^1.7.8",
         "path": "^0.12.7",
         "pg": "^8.0.3",
         "request": "^2.88.2"
